--- a/src/Elm/Parser.elm
+++ b/src/Elm/Parser.elm
@@ -10,13 +10,12 @@ module Elm.Parser exposing
 
 -}
 
-import Combine exposing (Parser, end)
 import Elm.Internal.RawFile as InternalRawFile
 import Elm.Parser.File exposing (file)
-import Elm.Parser.State exposing (State, emptyState)
 import Elm.RawFile exposing (RawFile)
 import Elm.Syntax.File exposing (File)
-import Parser exposing (DeadEnd)
+import Parser as Core exposing ((|.), DeadEnd)
+import ParserWithComments exposing (ParserWithComments)
 
 
 {-| **@deprecated** Use [`parseToFile`](#parseToFile) instead, which is simpler and doesn't require post-processing.
@@ -40,14 +39,4 @@ When parsing fails, the result will contain a list of errors indicating what wen
 -}
 parseToFile : String -> Result (List DeadEnd) File
 parseToFile input =
-    case Combine.runParser (withEnd file) emptyState input of
-        Ok ( _, fileContents ) ->
-            Ok fileContents
-
-        Err s ->
-            Err s
-
-
-withEnd : Parser State File -> Parser State File
-withEnd p =
-    p |> Combine.ignore end
+    Core.run (file |. Core.end) input