--- a/src/Elm/Parser/Comments.elm
+++ b/src/Elm/Parser/Comments.elm
@@ -3,8 +3,8 @@ module Elm.Parser.Comments exposing (declarationDocumentation, moduleDocumentati
 import Elm.Parser.Node as Node
 import Elm.Syntax.Documentation exposing (Documentation)
 import Elm.Syntax.Node exposing (Node)
-import Parser as Core exposing ((|.), Nestable(..))
-import ParserWithComments exposing (ParserWithComments)
+import Parser as Core exposing ((|.), Nestable(..), Parser)
+import ParserWithComments exposing (WithComments)
 import Rope
 
 
@@ -31,7 +31,7 @@ problemUnexpected =
     Core.problem "unexpected documentation comment"
 
 
-moduleDocumentation : ParserWithComments ()
+moduleDocumentation : Parser (WithComments ())
 moduleDocumentation =
     declarationDocumentation
         |> Core.map