--- a/src/Elm/Parser/Comments.elm
+++ b/src/Elm/Parser/Comments.elm
@@ -3,57 +3,57 @@ module Elm.Parser.Comments exposing (declarationDocumentation, moduleDocumentati
 import Elm.Parser.Node as Node
 import Elm.Syntax.Documentation exposing (Documentation)
 import Elm.Syntax.Node exposing (Node)
-import Parser as Core exposing ((|.), Nestable(..), Parser)
+import Parser exposing ((|.), Nestable(..), Parser)
 import ParserWithComments exposing (Comments)
 import Rope
 
 
-singleLineCommentCore : Core.Parser ()
+singleLineCommentCore : Parser.Parser ()
 singleLineCommentCore =
-    Core.symbol "--"
-        |. Core.chompWhile (\c -> c /= '\u{000D}' && c /= '\n')
+    Parser.symbol "--"
+        |. Parser.chompWhile (\c -> c /= '\u{000D}' && c /= '\n')
 
 
-multilineCommentString : Core.Parser String
+multilineCommentString : Parser.Parser String
 multilineCommentString =
-    Core.oneOf
-        [ Core.symbol "{-|"
-            |> Core.backtrackable
-            |> Core.map (\() -> problemUnexpected)
-        , Core.multiComment "{-" "-}" Nestable
-            |> Core.mapChompedString (\comment () -> Core.succeed comment)
+    Parser.oneOf
+        [ Parser.symbol "{-|"
+            |> Parser.backtrackable
+            |> Parser.map (\() -> problemUnexpected)
+        , Parser.multiComment "{-" "-}" Nestable
+            |> Parser.mapChompedString (\comment () -> Parser.succeed comment)
         ]
-        |> Core.andThen identity
+        |> Parser.andThen identity
 
 
-problemUnexpected : Core.Parser a
+problemUnexpected : Parser.Parser a
 problemUnexpected =
-    Core.problem "unexpected documentation comment"
+    Parser.problem "unexpected documentation comment"
 
 
 moduleDocumentation : Parser Comments
 moduleDocumentation =
     declarationDocumentation
-        |> Core.map (\comment -> Rope.one comment)
+        |> Parser.map (\comment -> Rope.one comment)
 
 
-declarationDocumentation : Core.Parser (Node Documentation)
+declarationDocumentation : Parser.Parser (Node Documentation)
 declarationDocumentation =
-    Core.oneOf
-        [ Core.oneOf
-            [ Core.symbol "{-|"
-                |> Core.backtrackable
-                |> Core.map (\() -> coreTemporaryProblem {- "falls back" to multiComment from the start -})
-            , Core.succeed (Core.succeed (Core.problem "not a documentation comment"))
+    Parser.oneOf
+        [ Parser.oneOf
+            [ Parser.symbol "{-|"
+                |> Parser.backtrackable
+                |> Parser.map (\() -> coreTemporaryProblem {- "falls back" to multiComment from the start -})
+            , Parser.succeed (Parser.succeed (Parser.problem "not a documentation comment"))
             ]
-            |> Core.andThen identity
-        , Core.multiComment "{-" "-}" Nestable
-            |> Core.mapChompedString (\comment () -> Core.succeed comment)
+            |> Parser.andThen identity
+        , Parser.multiComment "{-" "-}" Nestable
+            |> Parser.mapChompedString (\comment () -> Parser.succeed comment)
         ]
-        |> Core.andThen identity
+        |> Parser.andThen identity
         |> Node.parserCore
 
 
-coreTemporaryProblem : Core.Parser a
+coreTemporaryProblem : Parser.Parser a
 coreTemporaryProblem =
-    Core.problem ""
+    Parser.problem ""