--- a/src/Elm/Parser/Expose.elm
+++ b/src/Elm/Parser/Expose.elm
@@ -154,4 +154,5 @@ exposingVariants =
 
 functionExpose : Parser (WithComments TopLevelExpose)
 functionExpose =
-    ParserWithComments.fromCoreMap FunctionExpose Tokens.functionName
+    Core.map (\name -> { comments = Rope.empty, syntax = FunctionExpose name })
+        Tokens.functionName