--- a/src/ParserWithComments.elm
+++ b/src/ParserWithComments.elm
@@ -3,7 +3,6 @@ module ParserWithComments exposing
     , WithComments
     , many
     , manyWithoutReverse
-    , map
     , maybe
     , sepBy
     , sepBy1