… src/Combine.elm
… src/Elm/Parser/Expression.elm
… src/Elm/Parser/File.elm
… src/Elm/Parser/Imports.elm
… src/Elm/Parser/Layout.elm
… src/Elm/Parser/Modules.elm
… src/Elm/Parser/Node.elm
… src/Elm/Parser/Numbers.elm
… src/Elm/Parser/Patterns.elm
… src/Elm/Parser/State.elm
… src/Elm/Parser/Tokens.elm
… src/Elm/Parser/TypeAnnotation.elm
… src/Elm/Syntax/Node.elm
… src/Parser/Extra.elm
… src/Rope.elm
… tests/Elm/Parser/CaseExpressionTests.elm
… tests/Elm/Parser/CombineTestUtil.elm
… tests/Elm/Parser/CommentTest.elm
… tests/Elm/Parser/DeclarationsTests.elm
… tests/Elm/Parser/ExposeTests.elm
… tests/Elm/Parser/ExpressionTests.elm
… tests/Elm/Parser/FileTests.elm
… tests/Elm/Parser/GlslTests.elm
… tests/Elm/Parser/ImportsTests.elm
… tests/Elm/Parser/InfixTests.elm
… tests/Elm/Parser/LambdaExpressionTests.elm
… tests/Elm/Parser/LayoutTests.elm
… tests/Elm/Parser/LetExpressionTests.elm
… tests/Elm/Parser/ModuleTests.elm
… tests/Elm/Parser/ParserWithCommentsTestUtil.elm