as discussed in https://github.com/stil4m/elm-syntax/pull/233#issuecomment-2248836851

- before (taken from "after" of previous PR): 31419.200000000186 ms to do 100 runs => 314.1920000000019 ms/run
- after (purely combine→core): 27615.5 ms to do 100 runs => 276.155 ms/run
- after (+ additional optimizations): 18849.700000001118 ms to do 100 runs => 188.49700000001118 ms/run

So yeah, noticeable but not game-changing.

Code-style-wise, I personally like this new style a lot because of how obvious the information flow is. Though it is undeniably more verbose.