@jfmengels 2024-07-28
Well, this has been amazing once more. Another 50%+ performance increase :sweat_smile:
(though I wonder how it performs without eol2)

Great chances @lue-bird, thank you so much! I'm glad you started this experiment, and I hope you had fun! :wink: